Bump major version due to breaking changes and tidy up.
In the 1.x of the cookbook empty strings `''` were used to define unset
values. In 2.x we now use nil or undefined attributes to ensure we can
render an XML file which is fully idempotent and does not result in
Sonarr rewriting the file.

This commit also cleans up a few things:
 - Added copyright and contributor information.
 - Fixed cooksytle issues.
 - Fixed foodcritic issues.

Signed-off-by: Ryan Hass <rhass@users.noreply.github.com>